+++ packages/create-gasket-app/lib/scaffold/actions/create-hooks.js
@@ -1,26 +1,25 @@
-const createEngine = require('../create-engine');
-const action = require('../action-wrapper');
-const ConfigBuilder = require('../config-builder');
-const Files = require('../files');
+import action from '../action-wrapper.js';
+import { ConfigBuilder } from '../config-builder.js';
+import { Files } from '../files.js';
 
 /**
  * Executes the `create` hook for all registered plugins.
  * Adds `files` to context for plugins to add their files and templates.
  *
+ * @param {GasketEngine} gasket - Gasket API
  * @param {CreateContext} context - Create context
  * @returns {Promise} promise
  */
-async function createHooks(context) {
-  const { dest, presets = [], plugins = [], warnings } = context;
+async function createHooks({ gasket, context }) {
+  const { plugins = [], warnings } = context;
 
   const files = new Files();
-  const gasketConfig = ConfigBuilder.create({}, { orderBy: ['plugins'], warnings });
+  const gasketConfig = ConfigBuilder.create({}, { orderBy: ['plugins'], warnings, plugins });
   Object.assign(context, { files, gasketConfig });
 
-  const gasket = await createEngine({ dest, presets, plugins });
   await gasket.execApply('create', async function applyCreate(plugin, handler) {
     await handler(context.runWith(plugin));
   });
 }
 
-module.exports = action('Execute create hooks', createHooks);
+export default action('Execute create hooks', createHooks);